Make processes run under bear stoppable
If e.g. a long-running build is run under bear

    # bear -- make

and is stopped, the continued process dies since bear tries to handle `SIGCHLD`
even though all that happened was the child processes stopping and continuing.
Instead of trying to handle this case just don't handle it (since nothing bad
happened anyway).

This should fix #40 and make the fix in 88aa27 unnecessary.