Fix issue #51: execle causes bear to segfault on 32 bit systems.
Problem here is that we pass the va_list to bear_strings_build and then
use it again after returning.  Per the C99 standard that's not legal.
However, you can pass a pointer to the va_list, which is what we end up
doing.  See this stackoverflow question for quotes from the standard
and/or links:
http://stackoverflow.com/questions/3369588/pass-va-list-or-pointer-to-va-list