Fix shell_escape not escaping spaces
Prior to this change, If a command contained e.g. -Dtext=\"text\ with\ spaces\" or -Dtext="\"text with spaces\"", the function shell_escape did not correctly handle either case. It correctly escaped the quotes, but did not escape the spaces, resulting in it returning -D\"text with spaces\", when it should have been returning -D\"text\ with\ spaces\".