Don't cache the value of _NSGetEnviron on macOS (#290)
This causes an issue with xcrun because its purpose is to set SDKROOT,
which causes environ to change internally and when it calls execv (which
we interpose) we continue to use the old value we copied at launch time
rather than the new value. This is how environ has been defined
internally for just about the last decade so hopefully this shouldn't be
too poor of a workaround.

Also, fix a couple of typos that I happened upon.