Remove USE_SOURCE_PERMISSIONS from install command
The `USE_SOURCE_PERMISSIONS` option in the `install(DIRECTORY)` command
was causing issues with file permissions after `cmake install`.
Specifically, it was altering the permissions of the `bin`, `lib`, and
`share` folders, leading to `others` losing access to these directories.
Removing this option ensures that default permissions are maintained,
preventing unexpected access restrictions for users.